it-locale extension: codice fiscale
Person class is extended with CodiceFiscale() (fiscal code)
method. This method computes the fiscal code belonging to the
person. The method is stable, in the sense that multiple calls
on the same Person instance give the same code. The last part
of the code, however is fake, since importing the actual city
database would have been too heavy (~40kb).

Finance dataset has been extended, as well. The CodiceFiscale()
method in this case explicitly takes in input all person data
(firstname, lastname, birthday and gender) and computes the
corresponding code, allowing to get a fiscal code for a real
person.

The fiscal code is used in Italy to unambiguously identify individuals
residing in
Italy.